Move diagnostic check to scripts folder, rename to scripts/diagnostics.py
- Moved system_check.py to scripts/diagnostics.py
- Deleted hardware_test.py and system_check.py from root (layer 1 and 2 tests removed from this branch)
